+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## [Next]
+
+
 
 All notable changes to this project will be documented in this file.