+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## [Next]
 
+- Venus: Add support for Venus device types VNSA and VNSD
 - Fix loop when all configured devices have invalid/unknown device types.
 
 ## [1.5.2] - 2025-10-03